learn/audio/audio_streaming_speech_to_text/python/requirements.txt
requests
torch
torchaudio
websocket-client
